requirements.txt
streamlit>=1.40.0
faiss-cpu>=1.9.0
sentence-transformers>=3.0.0
groq>=0.12.0
pymupdf>=1.24.0
python-dotenv>=1.0.0
numpy>=1.26.0
fastapi>=0.110.0
uvicorn>=0.29.0